[CDS-35550]: Move manager side code from ArtifactType and move it down to 930-delegate-task
