+++ 400-rest/src/main/java/software/wings/service/impl/ServiceResourceServiceImpl.java
@@ -213,6 +213,7 @@ import software.wings.stencils.StencilCategory;
 import software.wings.stencils.StencilPostProcessor;
 import software.wings.utils.ApplicationManifestUtils;
 import software.wings.utils.ArtifactType;
+import software.wings.utils.artifacts.ArtifactCommandHelper;
 import software.wings.verification.CVConfiguration;
 
 import com.google.common.annotations.VisibleForTesting;
@@ -850,7 +851,7 @@ public class ServiceResourceServiceImpl implements ServiceResourceService, DataP
     if (appContainer != null && appContainer.getFamily() != null) {
       isInternal = appContainer.getFamily().isInternal();
     } else if (artifactType != null) {
-      isInternal = artifactType.isInternal();
+      isInternal = ArtifactCommandHelper.getArtifactCommands(artifactType).isInternal();
     }
     return isInternal;
   }
@@ -888,7 +889,7 @@ public class ServiceResourceServiceImpl implements ServiceResourceService, DataP
           }
           break;
         default:
-          commands = artifactType.getDefaultCommands();
+          commands = ArtifactCommandHelper.getArtifactCommands(artifactType).getDefaultCommands();
       }
     }
