+++ 400-rest/src/main/java/software/wings/utils/artifacts/AWSLambdaArtifactCommands.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright 2022 Harness Inc. All rights reserved.
+ * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
+ * that can be found in the licenses directory at the root of this repository, also available at
+ * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
+ */
+
+package software.wings.utils.artifacts;
+
+import static software.wings.beans.Graph.Builder.aGraph;
+import static software.wings.beans.Graph.graphIdGenerator;
+import static software.wings.beans.command.Command.Builder.aCommand;
+import static software.wings.sm.states.AwsLambdaState.AWS_LAMBDA_COMMAND_NAME;
+import static software.wings.utils.ArtifactType.AWS_LAMBDA;
+
+import static java.util.Arrays.asList;
+
+import software.wings.beans.GraphNode;
+import software.wings.beans.command.Command;
+import software.wings.beans.command.CommandType;
+
+import java.util.List;
+
+public class AWSLambdaArtifactCommands implements ArtifactCommands {
+  @Override
+  public boolean isInternal() {
+    return true;
+  }
+
+  @Override
+  public List<Command> getDefaultCommands() {
+    return asList(getAwsLambdaCommand());
+  }
+
+  private Command getAwsLambdaCommand() {
+    return aCommand()
+        .withCommandType(CommandType.INSTALL)
+        .withGraph(aGraph()
+                       .withGraphName(AWS_LAMBDA_COMMAND_NAME)
+                       .addNodes(GraphNode.builder()
+                                     .origin(true)
+                                     .id(graphIdGenerator("node"))
+                                     .name(AWS_LAMBDA_COMMAND_NAME)
+                                     .type(AWS_LAMBDA.name())
+                                     .build())
+                       .buildPipeline())
+        .build();
+  }
+}
