+++ 400-rest/src/main/java/software/wings/utils/artifacts/ArtifactCommands.java
@@ -1,0 +1,23 @@
+/*
+ * Copyright 2022 Harness Inc. All rights reserved.
+ * Use of this source code is governed by the PolyForm Free Trial 1.0.0 license
+ * that can be found in the licenses directory at the root of this repository, also available at
+ * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
+ */
+
+package software.wings.utils.artifacts;
+
+import software.wings.beans.command.Command;
+
+import java.util.List;
+
+public interface ArtifactCommands {
+  boolean isInternal();
+
+  /**
+   * Gets default commands.
+   *
+   * @return the default commands
+   */
+  List<Command> getDefaultCommands();
+}
